wifidebug: Add trace-cmd usage to start and stop gathering data.
When wifi debug mode is started trace-cmd start is executed with
the given buffer_size parameter.
Later when wifi debug mode is stopped, trace-cmd stop is executed
as well as trace-cmd extract to write the data to a file on disk.
In this way the steam client (or any dbus users...) can decide how
long to run trace-cmd for.
Change to create_dir_all so we don't get errors if it exists.

Signed-off-by: Jeremy Whiting <jeremy.whiting@collabora.com>